requirements.txt
litellm>=1.50.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.2.0
